snapcraft.yaml: switch to kernel and initrd plugin
The kernel and initrd plugins make building a kernel snap exceptionally
easy. As such, those plugins should be used for kernel snap builds.

Note that, currently, the change adding those plugins to snapcraft has
not yet been merged. This means that for anyone wanting to build kernel
snaps using this method, a snapcraft with that PR merged much be built
and installed.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>